Remove programming.el file
This filename didn't make any sense, and many of the modules contained
within the file were pertinent to not just programming but other
usecases for emacs. In the future I might reconsider the naming of
editor.el but at the moment no better name comes to mind. I also don't
want to create hundreds of files, one for each package/dependency.